+++ app/build.gradle
@@ -5,10 +5,10 @@ android {
     buildToolsVersion '25.0.2'
 
     defaultConfig {
-        applicationId "com.example.pcsaini779.hackathon"
+        applicationId "com.game.ajitrajeev.funwithsquare"
         minSdkVersion 16
         targetSdkVersion 25
-        versionCode 1
+        versionCode 2
         versionName "1.0"
     }
     buildTypes {
