+++ app/src/main/java/com/game/ajitrajeev/funwithsquare/Game.java
@@ -1,4 +1,4 @@
-package com.example.pcsaini779.hackathon;
+package com.game.ajitrajeev.funwithsquare;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
