+++ app/src/main/java/com/game/ajitrajeev/funwithsquare/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.pcsaini779.hackathon;
+package com.game.ajitrajeev.funwithsquare;
 
 import android.content.Intent;
 import android.graphics.Color;
