+++ app/src/main/java/com/game/ajitrajeev/funwithsquare/PatriciaTrie.java
@@ -1,4 +1,4 @@
-package com.example.pcsaini779.hackathon;
+package com.game.ajitrajeev.funwithsquare;
 
 /**
  * Created by jsrathore on 10/11/16.
