+++ app/src/main/java/com/game/ajitrajeev/funwithsquare/Team.java
@@ -1,4 +1,4 @@
-package com.example.pcsaini779.hackathon;
+package com.game.ajitrajeev.funwithsquare;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,33 +7,24 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
-import android.webkit.WebView;
 import android.widget.Button;
-import android.widget.LinearLayout;
-import android.widget.TextView;
 
-public class Instructions extends AppCompatActivity {
+public class Team extends AppCompatActivity {
 
     FloatingActionButton btnPlay;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_instructions);
-
-
+        setContentView(R.layout.activity_team);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         getSupportActionBar().setHomeButtonEnabled(true);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-
-
-        btnPlay = (FloatingActionButton) findViewById(R.id.btnPlayInst);
+        btnPlay = (FloatingActionButton) findViewById(R.id.btnPlayTeam);
         btnPlay.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(Instructions.this, Game.class));
+                startActivity(new Intent(Team.this, Game.class));
             }
         });
     }
