+++ app/src/main/res/layout/activity_game.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:ads="http://schemas.android.com/apk/res-auto"
     android:gravity="center_horizontal"
-    tools:context="com.example.pcsaini779.hackathon.Game"
+    tools:context="com.game.ajitrajeev.funwithsquare.Game"
     android:background="#00E676">
 
     <ScrollView
