+++ app/src/main/res/layout/activity_instructions.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.pcsaini779.hackathon.Instructions">
+    tools:context="com.game.ajitrajeev.funwithsquare.Instructions">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
