+++ app/src/main/res/layout/activity_team.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.pcsaini779.hackathon.Team">
+    tools:context="com.game.ajitrajeev.funwithsquare.Team">
 
     <android.support.design.widget.AppBarLayout
         android:layout_height="wrap_content"
