+++ app/src/main/res/layout/content_team.xml
@@ -12,7 +12,7 @@
     android:background="@color/colorPrimary"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_team"
-    tools:context="com.example.pcsaini779.hackathon.Team">
+    tools:context="com.game.ajitrajeev.funwithsquare.Team">
 
     <ScrollView
         android:layout_width="fill_parent"
