+++ app/src/main/res/menu/menu_instructions.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.pcsaini779.hackathon.Instructions">
+    tools:context="com.game.ajitrajeev.funwithsquare.Instructions">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
