+++ app/src/test/java/com/example/pcsaini779/hackathon/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.pcsaini779.hackathon;
+package com.game.ajitrajeev.funwithsquare;
 
 import org.junit.Test;
